Create Expedited Screening (No MAXIS)
This script will review an application for expedited status when the screener has no access to APPL cases in MAXIS.  For example, our support staff use the current script to screen for expedited, but cannot use it if there is no existing case.  This script allows them to use the expedited methodology even if there is no current MAXIS case.  It could be used to just determine expedited status, even before the CAF is filed.