Create Expedited Screening (No MAXIS)
I think I needed to correct my first pull request on this to update it to the correct Functions Library Block.  This script can be used by office support staff who cannot APPL cases in MAXIS to determine whether a case is expedited or not depdning on the CAF or verbal info given by the client.  This script will not create a case note nor will it check for a DISQ, but it does allow MAXIS inquiry users to do a basic screening for expedited.